requirements.txt
jupyter==1.1.1
notebook==7.3.2
pandas==2.2.3
Faker==36.1.1
datetime==5.5
mistralai==1.5.0 
streamlit==1.42.2
langchain_mistralai==0.2.7
openinference-instrumentation-mistralai==1.3.0
mistralai==1.5.1
arize-phoenix==8.12.1
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
openinference-instrumentation-langchain==0.1.35
langgraph==0.3.8